requirements.txt
apiclient==1.0.3
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
emoji==0.5.0
google-api-core==1.2.1
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
google-cloud-speech==0.34.0
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.13.0
httplib2==0.11.3
idna==2.7
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.1
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
PyAudio==0.2.11
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
rsa==3.4.2
selenium==3.14.0
six==1.11.0
SpeechRecognition==3.8.1
threadsafe-tkinter==1.0.1
twitter==1.18.0
uritemplate==3.0.0
urllib3==1.23
